hot_topic_agent/requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
pydantic
pandas
numpy
tiktoken
fastapi
uvicorn
jinja2